% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processingDistances.R
\name{addTrait}
\alias{addTrait}
\title{Add a new study with a trait name}
\usage{
addTrait(traits, df, name = "", traitName = "")
}
\arguments{
\item{traits}{A trait table that is used for conversion}

\item{df}{A dataframe of unique cg positions with traits}

\item{name}{Set empty by default. New study name}

\item{traitName}{Set empty by default. New trait name}
}
\value{
A dataframe with a new study appended
}
\description{
Add a new study with a trait name
}
