% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processingDistances.R
\name{appendMatrixSize}
\alias{appendMatrixSize}
\title{Append matrix rows}
\usage{
appendMatrixSize(col, data)
}
\arguments{
\item{col}{An array that has more positions that the matrix}

\item{data}{A dataframe which will be appended}
}
\value{
Return appended dataframe
}
\description{
Append matrix rows
}
\examples{
data <- appendMatrixSize(col, dataFrame)
}
