% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEwasData.R
\name{cleanAndMergeData}
\alias{cleanAndMergeData}
\title{Clean data and merge other matrices}
\usage{
cleanAndMergeData(data, merge_by = "Study.id")
}
\arguments{
\item{data}{A list of data matrices, first matrix must contain Study.id
and tissue}

\item{merge_by}{Parameter which indicated which column matrices will
be merged by}
}
\value{
Cleaned and merged dataframe file
}
\description{
Clean data and merge other matrices
}
\examples{
data <- cleanAndMergeData()
data <- cleanAndMergeData(data = list(matrix1, matrix2, matrix3))
data <- cleanAndMergeData( merge_by = "Study.id")
}
