% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processingDistances.R
\name{convertIDS}
\alias{convertIDS}
\title{Convert studyNames to ids for connections}
\usage{
convertIDS(df, cln)
}
\arguments{
\item{df}{A dataframe that contains 'from' and 'to' as numeric values}

\item{cln}{Either 'from' or 'to' colnames that are used for ID converion}
}
\value{
A dataframe with converted 'from' and 'to' columns
}
\description{
Convert studyNames to ids for connections
}
\examples{
df <-ConvertIDS(df, "from")
data <-ConvertIDS(df, "to")
}
