% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEwasData.R
\name{deleteCols}
\alias{deleteCols}
\title{Leave only certain columns}
\usage{
deleteCols(data)
}
\arguments{
\item{data}{A dataframe of merged matrices}
}
\value{
Dataframe with Study.id, Sample.size, Tissue, Trait, PMID, Ancestry
}
\description{
Leave only certain columns
}
\examples{
data <- deleteCols(dataFrame)
}
