% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseases.R
\name{dfForHistogram}
\alias{dfForHistogram}
\title{A dataframe for stacked histogram}
\usage{
dfForHistogram(data, positions = 0, traits = 0, anno = 0)
}
\arguments{
\item{data}{A data matrix with no modifications}

\item{positions}{An array of cg positions in new study}

\item{traits}{An array of trait names that was to be compared to.
By default set to 0 - compare to all}

\item{anno}{Annotation file for sorting probes by chromosomes and positions.
0 by default - not used}
}
\value{
A dataframe for visNetwork graph
}
\description{
A dataframe for stacked histogram
}
\examples{
df <- dfForHistogram(data, positions)
#' df <- dfForHistogram(data, positions, "Smoking", anno)
}
