% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processingDistances.R
\name{filterBy}
\alias{filterBy}
\title{Filter data according to given array}
\usage{
filterBy(data, colName, array)
}
\arguments{
\item{data}{A dataframe of merged data}

\item{colName}{A column name which will be filtered by}

\item{array}{An array that determins which rows will be left.}
}
\value{
A dataframe of filtered data
}
\description{
Filter data according to given array
}
\examples{
data <- filterBy(data, "Probe.id", c("cg02609880", "cg06874172"))
data <- filterBy(data, "Study.id", c("ES00034", "ES00035"))
data <- filterBy(data, "Trait", c("aging", "air pollution (PM2.5)"))
}
