% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processingDistances.R
\name{getEstimatedVals}
\alias{getEstimatedVals}
\title{Estimated values of odd ands pval}
\usage{
getEstimatedVals(data, col)
}
\arguments{
\item{data}{Matrix that has numerical values of True/False ratio}

\item{col}{An array which will be used as a base to test to}
}
\value{
Returns two matrices: first one is odds and the other - pvalues
}
\description{
Estimated values of odd ands pval
}
\examples{
list_of_datas <- getEstimatedVals(matrix, colnames)
}
