% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processingDistances.R
\name{getNewRow}
\alias{getNewRow}
\title{Get new row}
\usage{
getNewRow(df, data, col, name, shape, color)
}
\arguments{
\item{df}{A dataframe which will be appended}

\item{data}{A data matrix of TRUE/FALSE}

\item{col}{An array of cg positions that will be compared}

\item{name}{A name of the study}

\item{shape}{Shape for connection interpreting}

\item{color}{Color that is used in connections}
}
\value{
Returns a dataframe of unique positions and their traits
}
\description{
Get new row
}
\examples{
df <- getNewRow(df, data, positions, "new_study", "star", "purple")
}
