% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseases.R
\name{getSimilarTraits}
\alias{getSimilarTraits}
\title{Generate infromation from connection data}
\usage{
getSimilarTraits(data, study = 0)
}
\arguments{
\item{data}{A dataframe that was generated by makeConnections}

\item{study}{A single name or an array of study names that will be included}
}
\value{
Returns a dataframe with id, connecting from id, id's study name,
id's trait, connecting to id, id's study name, trait and overlap count
between them - matching cg positions count
}
\description{
Generate infromation from connection data
}
