% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processingDistances.R
\name{groupColor}
\alias{groupColor}
\title{Assign color by group}
\usage{
groupColor(data, colorData, colors)
}
\arguments{
\item{data}{A dataframe from makeConnections}

\item{colorData}{A dataframe with group names and study.id}

\item{colors}{A list of colors that should be assigned for each group}
}
\value{
A dataframe with colors by group name
}
\description{
Assign color by group
}
\examples{
colored_data <- groupColor(data, colorData, c("yellow", "green"))
colored_data <- groupColor(data, colorData, c("#22105", "#00FF4E"))
}
