% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseases.R
\name{hideArrows}
\alias{hideArrows}
\title{Leave only one arrow between same nodes}
\usage{
hideArrows(df, minLength = 0)
}
\arguments{
\item{df}{A dataframe from makeConnections function}

\item{minLength}{A number that defines the minimum count value. 0 by default}
}
\value{
A modified dataframe
}
\description{
Leave only one arrow between same nodes
}
\examples{
df <- hideArrows(df)
df <- hideArrows(df, 15)
}
