% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processingDistances.R
\name{makeArray}
\alias{makeArray}
\title{Make an array of TRUE/FALSE for given positions}
\usage{
makeArray(change, col)
}
\arguments{
\item{change}{An araay with names that will be changed into}

\item{col}{An array with names which will have same names
with same order as change array as well as TRUE/FALSE values}
}
\value{
Rearranged array
}
\description{
Make an array of TRUE/FALSE for given positions
}
\examples{
row <- rearrange(c("a", "b", "c"), c("c", "b", "t"))
}
