% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processingDistances.R
\name{makeConnections}
\alias{makeConnections}
\title{Make a dataframe of linked studies}
\usage{
makeConnections(
  data,
  minSize = 50,
  col = 15,
  name,
  type,
  traitMatrix,
  traits,
  color = c("red", "darkcyan"),
  shape = c("circle", "elipsis")
)
}
\arguments{
\item{data}{A data matrix with TRUE/FALSE values}

\item{minSize}{A number that defines minimum sample size: > X. 50 by default}

\item{col}{An array of cg positions in the new study}

\item{name}{A name of the study}

\item{type}{Disease type}

\item{traitMatrix}{A dataframe of studies and their traits}

\item{traits}{A list of trait names to filter by}

\item{color}{An array of colors to be used. First is for new and the
second if for other studies. Red and darcyan by default}

\item{shape}{An array of shapes for plot. Circle and elipsis by default}
}
\value{
A dataframe that is used in visNetwork
}
\description{
Make a dataframe of linked studies
}
