% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseases.R
\name{makeEdges}
\alias{makeEdges}
\title{Make edges for visNetwork}
\usage{
makeEdges(df, len)
}
\arguments{
\item{df}{A dataframe from function makeConnections()}

\item{len}{A number that defines the minimum length for each edge}
}
\value{
A new dataframe with edges for visNetwork
}
\description{
Make edges for visNetwork
}
