% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseases.R
\name{makeNodes}
\alias{makeNodes}
\title{Make nodes for visNetwork}
\usage{
makeNodes(df)
}
\arguments{
\item{df}{A dataframe from function makeConnections}
}
\value{
A new dataframe of nodes that will be used in visNetwork
}
\description{
Make nodes for visNetwork
}
