% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processingDistances.R
\name{maxMatrix}
\alias{maxMatrix}
\title{Append all missing cg positions from annotation}
\usage{
maxMatrix(data)
}
\arguments{
\item{data}{Matrix of positions and studies with TRUE/FALSE}
}
\value{
Data matrix with new positions at the end
}
\description{
Append all missing cg positions from annotation
}
\examples{
data <- maxMatrix(dataMatrix)
}
