% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEwasData.R
\name{mergeData}
\alias{mergeData}
\title{Merge several data matrices}
\usage{
mergeData(matrices, merge_by)
}
\arguments{
\item{matrices}{A list of matrices to merge}

\item{merge_by}{A column name that will be merged by.
Must have same column name}
}
\value{
Returns merged dataframe of listed matrices
}
\description{
Merge several data matrices
}
\examples{
data<- mergeData(list(mat1, mat2, mat3), "Study.id")
}
