% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processingDistances.R
\name{processData}
\alias{processData}
\title{Process data into a dataframe for connections}
\usage{
processData(
  data,
  positions = 0,
  studies = 0,
  studByTraits = 0,
  maxPositions = FALSE,
  minSampleSize = 100
)
}
\arguments{
\item{data}{A dataframe of general data}

\item{positions}{An array of cg positions in the new study
that wants to be checked, 0 by default - all positions included}

\item{studies}{An array of study names that wants to be checked,
0 by default - all study names included}

\item{studByTraits}{A list of trait names to be filtered by. 0 by default}

\item{maxPositions}{Set to FALSE by Default. If set to TRUE,
900k positions will be used in general}

\item{minSampleSize}{Minimal sample size to be used. Set to 100 by default}
}
\value{
A matrix of positions x studies with TRUE/FALSE values. If a cg
position is found in specific study - valuse set to TRUE
}
\description{
Process data into a dataframe for connections
}
\examples{
data1 <- processData(data, col)
data1 <- processData(data, col, positions, studies)
data1 <- processData(data, col, positions, minSampleSize = 500)
}
