% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEwasData.R
\name{rawMockData}
\alias{rawMockData}
\title{Get Raw Mock Data}
\usage{
rawMockData()
}
\value{
List of matrices: cohorts, associations, studies and annotations
}
\description{
Get Raw Mock Data
}
\examples{
data <- rawMockData()
}
