% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEwasData.R
\name{removeDuplicates}
\alias{removeDuplicates}
\title{Remove duplicate Study.id}
\usage{
removeDuplicates(data)
}
\arguments{
\item{data}{Data matrix that must have Study.id, Platform, Sample.size
Tissue and Ancestry columns}
}
\value{
Returns new matrix that contains data with no duplicated Study.id
}
\description{
Remove duplicate Study.id
}
\examples{
newData <- removeDuplicates(data)
}
