% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processingDistances.R
\name{rowAndCol}
\alias{rowAndCol}
\title{True and False matrix}
\usage{
rowAndCol(row, col)
}
\arguments{
\item{row}{An array of rowanames from the matrix}

\item{col}{An array of colnames from the matrix}
}
\value{
Returns a matrix of rows and columns. Value is either TRUE or
FALSE depending on column and row match
}
\description{
True and False matrix
}
\examples{
matrix <- rowAndCwol(c("cg00000000", "cg00000000"), c("Headache","Smoking"))
matrix <- rowAndCol(dataFrame$Probe, dataFrame$Trait)
}
