% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseases.R
\name{stackedHistogram}
\alias{stackedHistogram}
\title{A stacked histogram}
\usage{
stackedHistogram(df, minCount = 1, path = 0)
}
\arguments{
\item{df}{A dataframe from function dfForHistrogram}

\item{minCount}{A number that will limit position count in traits for show}

\item{path}{A path to save file to. 0 by default}
}
\value{
A highcharter stacked histogram
}
\description{
A stacked histogram
}
\examples{
hist <- stackedHistogram(df)
hist <- stackedHistogram(df, 0, path = 'C:/.../histogram.html')
}
