% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processingDistances.R
\name{traitTable}
\alias{traitTable}
\title{Trait table}
\usage{
traitTable(data)
}
\arguments{
\item{data}{Primary dataframe that is used for identifying
traits according to study names}
}
\value{
A dataframe with trait for each study name
}
\description{
Trait table
}
