% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processingDistances.R
\name{appendMatrixSize}
\alias{appendMatrixSize}
\title{Append matrix rows}
\usage{
appendMatrixSize(col, data)
}
\arguments{
\item{col}{A vector that has different positions than the matrix}

\item{data}{A dataframe which will be appended}
}
\value{
Return appended dataframe
}
\description{
Append matrix rows
}
\examples{
data <- appendMatrixSize(col = c("cg11001140", "cg88888888"), dataFrame)
}
