% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseases.R
\name{getVisNetwork}
\alias{getVisNetwork}
\title{Get visNetwork plot}
\usage{
getVisNetwork(data, edgeLen = 100, path = 0)
}
\arguments{
\item{data}{A dataframe that was generated by makeConnections}

\item{edgeLen}{A number that defines minimum edge length.
100 by default}

\item{path}{A path to a folder where the graph should be saved.
0 by default that indicates it wil not be saved. Must have .html extension}
}
\value{
A visNetwork graph
}
\description{
Get visNetwork plot
}
\examples{
network <- getVisNetwork(data)
network <- getVisNetwork(data, 'C:/../home/test.html')
}
